problem: jeromq is not thread-safe
Solution: port CLIENT and SERVER sockets from libzmq, which are thread-safe sockets.

The thread-safe sockets family includes other sockets, which are not part of this commit.

Also, CLIENT and SERVER cannot be used with a poller at the moment. For poller support, zmq_poller from libzmq has to be ported as well.